+++ src/components/FormElements/Select/Select.tsx
@@ -1,43 +1,17 @@
-// Inspired by this https://codesandbox.io/embed/m75wlyx3oy
-import React, {
-  ForwardRefRenderFunction,
-  PropsWithChildren,
-  forwardRef,
-  isValidElement,
-  useRef,
-  useState,
-} from "react";
+import React, { ForwardRefRenderFunction, forwardRef, isValidElement, useRef } from "react";
 import classNames from "classnames";
-import {
-  ActionMeta,
-  DropdownIndicatorProps,
-  GroupBase,
-  MultiValue,
-  SelectInstance,
-  SingleValue,
-  ValueContainerProps,
-  components,
-} from "react-select";
+import { ActionMeta, MultiValue, SelectInstance, SingleValue } from "react-select";
 import { SerializedStyles } from "@emotion/react";
-import { useClickAway } from "ahooks";
 import { AddOperatorSVG, InfoCircledSVG } from "../../../icons";
 import Label from "../Label/Label";
 import Tooltip from "../../Tooltip/Tooltip";
-import CustomValueContainer from "./components/CustomValueContainer";
-import CustomOptionComponent from "./components/CustomOption";
-import CustomMenuList from "./components/CustomMenuList";
-import CustomSingleValue from "./components/CustomSingleValue";
-import CustomMultiValueLabel from "./components/CustomMultiValueLabel";
 import { resolveStyles, selectContainer } from "./styles";
 import { CustomOption, CustomSelectProps } from "./types";
-import {
-  MAX_MENU_HEIGHT,
-  INNER_PLACEHOLDER,
-  MIN_WIDTH,
-  MAX_WIDTH,
-  OUTER_PLACEHOLDER,
-} from "./constants";
+import { MAX_MENU_HEIGHT, MIN_WIDTH, MAX_WIDTH, PLACEHOLDER } from "./constants";
 import { containerClassNames, renderSelect } from "./helpers";
+import CustomMultiValueLabel from "./components/CustomMultiValueLabel";
+import CustomSingleValue from "./components/CustomSingleValue";
+import CustomOptionComponent from "./components/CustomOptionComponent";
 
 const Select: ForwardRefRenderFunction<
   SelectInstance<CustomOption>,
@@ -53,30 +27,24 @@ const Select: ForwardRefRenderFunction<
     inline = false,
     status = "valid",
     isInlineFlex = false,
-    hasInnerSearch = false,
     isMulti = false,
     creatableTooltip = "Create",
     maxMenuHeight = MAX_MENU_HEIGHT,
-    menuIsOpen,
-    innerPlaceholder = INNER_PLACEHOLDER,
     minWidth = MIN_WIDTH,
     maxWidth = MAX_WIDTH,
-    placeholder: outerPlaceholder = OUTER_PLACEHOLDER,
+    placeholder = PLACEHOLDER,
     tooltipContent = "",
-    countOptionsForInnerSearch = 10,
-    onChange,
-    isInputValid,
-    checkIfInputIsSelected,
-    closeMenuOnSelect,
+    minNumberOfOptionsToEnableSearch = 10,
+    isValidNewOption,
     menuMaxWidth,
+    onChange,
+    loadOptions,
+    isSearchable,
     ...rest
   } = props;
+
   const hasLabel = Boolean(label);
-
   const containerRef = useRef<HTMLInputElement>(null);
-
-  const [isFocused, setIsFocused] = useState(false);
-  const [inputValue, setInputValue] = useState("");
   const labelClassname = classNames({
     required,
   });
@@ -88,124 +56,68 @@ const Select: ForwardRefRenderFunction<
     }, 0);
   };
 
-  const shouldShowInnerSearch = () => {
-    // Force show inner search if the number of options exceeds 10 or certain conditions are met
-    const isAsyncType = type === "async";
-    const hasManyOptions = countOptions() > countOptionsForInnerSearch;
+  const isSelectSearchable = () => {
+    if (isSearchable !== undefined) {
+      return isSearchable;
+    }
 
-    return isAsyncType || hasManyOptions || hasInnerSearch;
+    const isAsyncType = type === "async";
+    const hasManyOptions = countOptions() > minNumberOfOptionsToEnableSearch;
+    return isAsyncType || hasManyOptions;
   };
 
-  const innerSearchEnabled = shouldShowInnerSearch();
   const shouldRenderTooltip =
     (tooltipContent && typeof tooltipContent === "string" && tooltipContent !== "") ||
     isValidElement(tooltipContent);
 
-  const styles = resolveStyles({ size, hasInnerSearch, menuMaxWidth });
+  const styles = resolveStyles({ size, menuMaxWidth });
 
   const formatCreateLabel = (inputValue: string) => (
-    <div>
-      <Tooltip content={creatableTooltip}>
-        <div className="select-create-label">
-          <span>{inputValue} </span>
-          <span>
-            <AddOperatorSVG color="black" height={12} />
-          </span>
-        </div>
-      </Tooltip>
-    </div>
+    <Tooltip content={creatableTooltip} parentProps={{ "aria-label": inputValue }}>
+      <div className="select-create-label">
+        <span>{inputValue} </span>
+        <span>
+          <AddOperatorSVG color="black" height={12} />
+        </span>
+      </div>
+    </Tooltip>
   );
-
-  const CustomDropdownIndicator: React.FC<DropdownIndicatorProps<CustomOption, boolean>> = (
-    props,
-  ) => {
-    const {
-      selectProps: { menuIsOpen },
-    } = props;
-
-    const handleCloseMenu = () => {
-      if (menuIsOpen) {
-        setIsFocused(false);
-      }
-    };
-
-    return (
-      <components.DropdownIndicator {...props}>
-        <components.DownChevron onClick={handleCloseMenu} />
-      </components.DropdownIndicator>
-    );
-  };
 
   const customSelectProps = {
     ...rest,
+    id: id,
+    "aria-label": id,
     ref: forwardedRef,
     styles,
-    blurInputOnSelect: closeMenuOnSelect || !isMulti,
-    closeMenuOnSelect: closeMenuOnSelect || !isMulti,
     isMulti,
     classNames: {
-      control: ({ isFocused }: { isFocused: boolean }) =>
-        containerClassNames(status, size, isFocused),
+      control: () => containerClassNames(status, size),
       option: ({ isSelected }: { isSelected: boolean }) =>
         `${isSelected ? "selected" : ""} option-${size}`,
     },
     components: {
       IndicatorSeparator: () => null,
-      DropdownIndicator: (
-        props: PropsWithChildren<
-          DropdownIndicatorProps<CustomOption, boolean, GroupBase<CustomOption>>
-        >,
-      ) => CustomDropdownIndicator(props),
-      MenuList: CustomMenuList,
-      ValueContainer: (
-        props: PropsWithChildren<
-          ValueContainerProps<CustomOption, boolean, GroupBase<CustomOption>>
-        >,
-      ) => CustomValueContainer({ ...props, isFocused }),
       Option: CustomOptionComponent,
       SingleValue: CustomSingleValue,
       MultiValueLabel: CustomMultiValueLabel,
     },
     formatCreateLabel,
-    isSearchable: false,
+    isSearchable: isSelectSearchable(),
     maxMenuHeight,
-    menuIsOpen: menuIsOpen ?? (isFocused || undefined), // when menuIsOpen is true, dropdown will stay open
     options,
-    placeholder: outerPlaceholder,
-    inputValue: inputValue,
-    onMenuInputFocus: () => setIsFocused(true),
-    onMouseDown: (e: MouseEvent) => e.stopPropagation(),
-    innerPlaceholder,
-    hasInnerSearch: innerSearchEnabled,
+    placeholder,
     type,
+    isValidNewOption,
+    loadOptions,
+    onMouseDown: (e: MouseEvent) => e.stopPropagation(),
     onChange: (
       option: MultiValue<CustomOption> | SingleValue<CustomOption>,
       action: ActionMeta<CustomOption>,
     ) => {
-      setIsFocused(false);
-      onChange && onChange(option, action);
+      onChange?.(option, action);
     },
-    onInputChange: (val: string) => setInputValue(val),
-    isValidNewOption: isInputValid,
-    noOptionsMessage: () => checkIfInputIsSelected && checkIfInputIsSelected(inputValue),
     isOptionDisabled: (option: CustomOption): boolean => Boolean(option.disabled),
   };
-
-  useClickAway(
-    (e) => {
-      // Ignore clicks on the close icon, can be one of the three following:
-      const { nodeName, className } = e.target as HTMLElement;
-      const isSvg = nodeName === "svg";
-      const isPath = nodeName === "path";
-      const isCloseIcon = className === "close-icon";
-
-      if (isSvg || isPath || isCloseIcon) return;
-      setIsFocused(false);
-      setInputValue("");
-    },
-
-    [containerRef],
-  );
 
   return (
     <div
@@ -216,7 +128,6 @@ const Select: ForwardRefRenderFunction<
           isInlineFlex,
           minWidth,
           maxWidth,
-          hasInnerSearch,
         })
       }
     >
@@ -232,7 +143,7 @@ const Select: ForwardRefRenderFunction<
           </Label>
           {shouldRenderTooltip && (
             <div data-testid={`${id}-tooltip`}>
-              <Tooltip content={tooltipContent}>
+              <Tooltip content={tooltipContent} parentProps={{ "aria-label": label }}>
                 <InfoCircledSVG height={20} />
               </Tooltip>
             </div>