+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+## [5.30.4](https://github.com/epignosis/gnosis/compare/v5.30.3...v5.30.4) (2024-09-30)
+
+## [5.30.3](https://github.com/epignosis/gnosis/compare/v5.30.2...v5.30.3) (2024-09-30)
+
+
+### Bug Fixes
+
+* **a11y toggle description color:** description color changes ([#415](https://github.com/epignosis/gnosis/issues/415)) ([1531c0c](https://github.com/epignosis/gnosis/commit/1531c0c22c5461d8644eb21bd6e2b8082743aa3a))
+
+## [5.30.2](https://github.com/epignosis/gnosis/compare/v5.30.1...v5.30.2) (2024-09-26)
+
+## [5.30.1](https://github.com/epignosis/gnosis/compare/v5.30.0...v5.30.1) (2024-09-26)
+
+# [5.30.0](https://github.com/epignosis/gnosis/compare/v5.29.5...v5.30.0) (2024-09-26)
+
+
+### Features
+
+* **select:** props change ([#411](https://github.com/epignosis/gnosis/issues/411)) ([50be079](https://github.com/epignosis/gnosis/commit/50be079f46b07f5a3bda2589a0d4ee16eb902429))
+
+## [5.29.5](https://github.com/epignosis/gnosis/compare/v5.29.4...v5.29.5) (2024-09-25)
+
+## [5.29.4](https://github.com/epignosis/gnosis/compare/v5.29.3...v5.29.4) (2024-09-25)
+
+## [5.29.3](https://github.com/epignosis/gnosis/compare/v5.29.2...v5.29.3) (2024-09-25)
+
+## [5.29.2](https://github.com/epignosis/gnosis/compare/v5.29.1...v5.29.2) (2024-09-25)
+
+## [5.29.1](https://github.com/epignosis/gnosis/compare/v5.29.0...v5.29.1) (2024-09-24)
+
+# [5.29.0](https://github.com/epignosis/gnosis/compare/v5.28.6...v5.29.0) (2024-09-23)
+
+
+### Features
+
+* add tooltipIcon to Input and tooltipContent to DropdownItem ([#403](https://github.com/epignosis/gnosis/issues/403)) ([b0debd8](https://github.com/epignosis/gnosis/commit/b0debd88b28d3ea4bbb00afb5274594fb16733c3))
+
+## [5.28.6](https://github.com/epignosis/gnosis/compare/v5.28.5...v5.28.6) (2024-09-23)
+
+
+### Bug Fixes
+
+* **a11y fix tooltip and aria label error:** a11y fix tooltip aria label ([#405](https://github.com/epignosis/gnosis/issues/405)) ([02966ee](https://github.com/epignosis/gnosis/commit/02966ee2d345211aba952cb9b1f84712d0e510df))
+
+## [5.28.5](https://github.com/epignosis/gnosis/compare/v5.28.4...v5.28.5) (2024-09-19)
+
+
+### Bug Fixes
+
+* mistake on release ([#402](https://github.com/epignosis/gnosis/issues/402)) ([ed874e4](https://github.com/epignosis/gnosis/commit/ed874e410e42911191a76d7bb87c0ae016737526))
+
 ## [5.28.4](https://github.com/epignosis/gnosis/compare/v5.28.3...v5.28.4) (2024-09-12)
 
 ## [5.28.3](https://github.com/epignosis/gnosis/compare/v5.28.2...v5.28.3) (2024-09-06)